requirements.txt
openai>=0
gradio>=0
python-dotenv>=0